change range to u32 (#15)
* change range to u32

* remove conversions

+++ cas_types/src/lib.rs
@@ -14,8 +14,8 @@ pub struct UploadXorbResponse {
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Range {
-    pub start: u64,
-    pub end: u64,
+    pub start: u32,
+    pub end: u32,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
